// src/shared/api/models/getApiStatisticsAppointmentsByTimeDateFilterType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiStatisticsAppointmentsByTimeDateFilterType =
    (typeof GetApiStatisticsAppointmentsByTimeDateFilterType)[keyof typeof GetApiStatisticsAppointmentsByTimeDateFilterType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiStatisticsAppointmentsByTimeDateFilterType = {
    bookingDate: 'bookingDate',
    createdDate: 'createdDate',
} as const;