// src/shared/api/models/getApiStatisticsAppointmentsByTimeGroupBy.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiStatisticsAppointmentsByTimeGroupBy =
    (typeof GetApiStatisticsAppointmentsByTimeGroupBy)[keyof typeof GetApiStatisticsAppointmentsByTimeGroupBy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiStatisticsAppointmentsByTimeGroupBy = {
    time: 'time',
    week: 'week',
} as const;