// src/shared/api/models/getApiStatisticsAppointmentsByTimeSource.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiStatisticsAppointmentsByTimeSource =
    (typeof GetApiStatisticsAppointmentsByTimeSource)[keyof typeof GetApiStatisticsAppointmentsByTimeSource];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiStatisticsAppointmentsByTimeSource = {
    STOREAPP: 'STOREAPP',
    DIRECTWEBSTORE: 'DIRECTWEBSTORE',
    ALL: 'ALL',
} as const;